Fixing file size issue
git-svn-id: svn://192.168.1.10/public/Phile-Backups@25397 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ schema.sql
@@ -32,7 +32,7 @@ CREATE TABLE IF NOT EXISTS `backup_files` (
     `backup_id`    INT          NOT NULL,
     `filename`     VARCHAR(255) NOT NULL,
     `filepath`     TEXT         NOT NULL,
-    `filesize`     INT          DEFAULT 0,
+    `filesize`     BIGINT       DEFAULT 0,
     `file_mtime`   INT          DEFAULT 0,
     `status`       VARCHAR(20)  DEFAULT 'active',
     `discovered_at` INT         DEFAULT 0,
